played around with different errorhandling

+++ src/main/java/com/example/users/service/UserServiceImpl.java
@@ -4,10 +4,9 @@ import com.example.users.controller.UserRepository;
 import com.example.users.model.User;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
 import javax.transaction.Transactional;
-import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class UserServiceImpl implements UserService {
@@ -23,8 +22,13 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public User getUserById(Integer id) {
-        return repo.findById(id).get();
-        // FEL HANTERING
+
+        Optional<User> user = repo.findById(id);
+        if(user.isPresent())
+            return user.get();
+        else return null;  // vill istället kasta ett exception här för att slippa null-värde,
+        // eller flytta ut felhantering utanför metoden,
+        // men känns som man vill ha error-handling i denna class ?
     }
 
     @Override
@@ -34,7 +38,13 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public void deleteUser(User user) {
-        repo.delete(user);
+
+            if(repo.existsById(user.getId())){
+                repo.delete(user);
+            } else {
+                System.out.println("Vi hittar ingen med id " + user.getId() +" i vår databas");
+            }
+
     }
 
     @Override
